+++ UltitemsCyan/Items/Lunar/SilverThread.cs
@@ -1,8 +1,6 @@
 ﻿using BepInEx.Configuration;
-using Mono.Cecil.Cil;
-using MonoMod.Cil;
 using RoR2;
-using System;
+using static System.Math;
 using System.Collections.Generic;
 using UltitemsCyan.Items.Untiered;
 using UnityEngine;
@@ -54,107 +52,112 @@ namespace UltitemsCyan.Items.Lunar
         protected override void Hooks()
         {
             // Gain additional items
-            On.RoR2.Inventory.GiveItem_ItemIndex_int += Inventory_GiveItem_ItemIndex_int;
+            //On.RoR2.Inventory.GiveItem_ItemIndex_int += Inventory_GiveItem_ItemIndex_int;
+            On.RoR2.Inventory.GiveItemPermanent_ItemIndex_int += Inventory_GiveItemPermanent_ItemIndex_int;
+            On.RoR2.Inventory.GiveItemTemp += Inventory_GiveItemTemp;
 
             // Chance of Death
-            IL.RoR2.HealthComponent.TakeDamageProcess += HealthComponent_TakeDamageProcess;
+            //IL.RoR2.HealthComponent.TakeDamageProcess += HealthComponent_TakeDamageProcess;
+            On.RoR2.HealthComponent.SendDamageDealt += HealthComponent_SendDamageDealt;
 
             // Increase cauldron and 3D printer cost
             On.RoR2.PurchaseInteraction.OnInteractionBegin += PurchaseInteraction_OnInteractionBegin;
             // Increase scrapper cost
-            On.RoR2.ScrapperController.BeginScrapping_UniquePickup += ScrapperController_BeginScrapping_UniquePickup;
+            //On.RoR2.ScrapperController.BeginScrapping_UniquePickup += ScrapperController_BeginScrapping_UniquePickup;
             //On.RoR2.ScrapperController.BeginScrapping += ScrapperController_BeginScrapping; // Old Scrapper Method
             // Remove Item on Death
             On.RoR2.CharacterBody.OnDeathStart += CharacterBody_OnDeathStart;
         }
-
         private int MaxStack(Inventory inv)
-        {//Test This Code
-            return Math.Min(inv.GetItemCountEffective(item), maxStack);
+        {
+            return Min(inv.GetItemCountEffective(item), maxStack);
         }
 
-        // Chance of Death on hit
-        private void HealthComponent_TakeDamageProcess(ILContext il)
+        // Return true if Silver Thread should give an additional item
+        private bool CheckSilverThreadRoll(Inventory inventory, ItemIndex itemIndex)
         {
-            ILCursor c = new(il); // Make new ILContext
-
-            int num14 = -1;
-
-            //Log.Warning("Silver Thread Take Damage");
-
-            // Inject code just before damage is subtracted from health
-            // Go just before the "if (num14 > 0f && this.barrier > 0f)" line, which is equal to the following instructions
-
-            // 1170 ldloc.s V_49 (49)             // Previous For loop k value
-            if (c.TryGotoNext(MoveType.Before,                        // 1673 callvirt remove Buff... // Previous For Loop branch
-                x => x.MatchLdloc(out num14),                         // 1674 ldloc.s V_8 (8)
-                x => x.MatchLdcR4(0f),                                // 1675 ldc.r4 0
-                x => x.Match(OpCodes.Ble_Un_S),                       // 1676 ble.un.s 1200 (0DE8) ldloc.s V_8 (8)
-                x => x.MatchLdarg(0),                                 // 1677 ldarg.0
-                x => x.MatchLdfld<HealthComponent>("barrier"),        // 1678 ldfld float32 RoR2.HealthComponent::barrier
-                x => x.MatchLdcR4(0f),                                // 1679 ldc.r4 0
-                x => x.Match(OpCodes.Ble_Un_S))                       // 1680 ble.un.s 1200 (0DE8) ldloc.s V_8 (8)
-            )
+            Log.Debug(" ??? ??? SilverThread Checking Rolls");
+            if (itemIndex == 0 || !ItemCatalog.GetItemDef(itemIndex))
             {
-
-                //Log.Debug(" * * * Start C Index: " + c.Index + " > " + c.ToString());
-                // [Warning:UltitemsCyan] * **Start C Index: 1173 > // ILCursor: System.Void DMD<RoR2.HealthComponent::TakeDamage>?-822050560::RoR2.HealthComponent::TakeDamage(RoR2.HealthComponent,RoR2.DamageInfo), 1173, Next
-                // IL_0e8a: blt.s IL_0e70
-                // IL_0e8f: ldloc.s V_7
-
-                c.Index++;
-
-
-                //Log.Debug(" * * * +1 Working Index: " + c.Index + " > " + c.ToString());
-                // [Warning:UltitemsCyan]  * * * Working Index: 1174 > // ILCursor: System.Void DMD<RoR2.HealthComponent::TakeDamage>?-822050560::RoR2.HealthComponent::TakeDamage(RoR2.HealthComponent,RoR2.DamageInfo), 1174, None
-                // IL_0e8f: ldloc.s V_7
-                // IL_0e91: ldc.r4 0
-
-                //c.GotoNext(MoveType.Before, x => x.MatchLdcR4(0f));
-                //Log.Warning(" * * * Before LdcR4 0f: " + c.Index + " > " + c.ToString());
-                // [Warning:UltitemsCyan]  * * * Before LdcR4 0f: 1174 > // ILCursor: System.Void DMD<RoR2.HealthComponent::TakeDamage>?-822050560::RoR2.HealthComponent::TakeDamage(RoR2.HealthComponent,RoR2.DamageInfo), 1174, Next
-                // IL_0e8f: ldloc.s V_7
-                // IL_0e91: ldc.r4 0
-
-                _ = c.Emit(OpCodes.Ldarg, 0);       // Load Health Component
-                _ = c.Emit(OpCodes.Ldloc, 1);       // Load Attacker Character Body
-                _ = c.Emit(OpCodes.Ldloc, num14);   // Load Total Damage
-
-                // Run custom code
-                _ = c.EmitDelegate<Action<HealthComponent, CharacterBody, float>>((hc, aCb, td) =>
+                Log.Debug("SilverThread found either none or impossible item? Index: " + itemIndex);
+            }
+            else if (NetworkServer.active && inventory && ItemCatalog.GetItemDef(itemIndex).tier != ItemTier.NoTier && itemIndex != item.itemIndex)
+            {
+                // Precaution incase something causes an infinity loop of items
+                //inSilverAlready = true;
+                //Log.Debug("Do you have silver?");
+                int grabCount = MaxStack(inventory);
+                if (grabCount > 0)
                 {
-                    CharacterBody cb = hc.body;
-                    //Log.Debug("Health: " + hc.fullCombinedHealth + "\t Body: " + cb.GetUserName() + "\t Damage: " + td);
-                    //Log.Warning("Damage Info " + di.ToString() + " with " + di.damage + " initial damage");
-                    if (cb && cb.master && cb.master.inventory && aCb)
+                    //Log.Debug("yes I do");
+                    Log.Debug("Thread Chance: " + (baseThreadChance + stackThreadChance * (grabCount - 1)));
+                    if (Util.CheckRoll(baseThreadChance + stackThreadChance * (grabCount - 1)))
                     {
-                        //if (aCb.master.inventory){ Log.Debug("and has inventory"); }
-                        int grabCount = MaxStack(cb.master.inventory);
-                        if (grabCount > 0)
-                        {
-                            Log.Debug(cb.GetUserName() + " takes " + td + "\t silver thread damage with " + hc.fullCombinedHealth + "\t health");
-                            float deathChance = td / hc.fullCombinedHealth * 100f * grabCount;
-                            if (deathChance > 100f)
-                            {
-                                deathChance = 100f;
-                            }
-                            Log.Debug("Chance of Snapping: " + deathChance);
-                            if (Util.CheckRoll(deathChance))
-                            {
-                                SnapBody(cb, aCb);
-                            }
-                        }
+                        Log.Debug("Extra silver: " + ItemCatalog.GetItemDef(itemIndex).name);
+                        // TODO add effect
+                        return true;
                     }
-                });
-                //Log.Debug(il.ToString());
+                }
+            }
+            return false;
+        }
+
+        private void Inventory_GiveItemTemp(On.RoR2.Inventory.orig_GiveItemTemp orig, Inventory self, ItemIndex itemIndex, float countToAdd)
+        {
+            Log.Debug("SilverThread got Temporary Item");
+            if (countToAdd == 1f && CheckSilverThreadRoll(self, itemIndex))
+            {
+                orig(self, itemIndex, countToAdd + 1f);
             }
             else
             {
-                Log.Warning("Silver cannot find '(num14 > 0f && this.barrier > 0f)'");
+                orig(self, itemIndex, countToAdd);
             }
         }
 
-        // Kill character body
+        private void Inventory_GiveItemPermanent_ItemIndex_int(On.RoR2.Inventory.orig_GiveItemPermanent_ItemIndex_int orig, Inventory self, ItemIndex itemIndex, int countToAdd)
+        {
+            Log.Debug("SilverThread got Permanent Permanent Item");
+            if (countToAdd == 1 && CheckSilverThreadRoll(self, itemIndex))
+            {
+                orig(self, itemIndex, countToAdd + 1);
+            }
+            else
+            {
+                orig(self, itemIndex, countToAdd);
+            }
+        }
+
+        private void HealthComponent_SendDamageDealt(On.RoR2.HealthComponent.orig_SendDamageDealt orig, DamageReport damageReport)
+        {
+            orig(damageReport);
+
+            CharacterBody attackerBody = damageReport.attackerBody;
+            CharacterBody victimBody = damageReport.victimBody;
+            //Log.Debug("Health: " + hc.fullCombinedHealth + "\t Body: " + cb.GetUserName() + "\t Damage: " + td);
+            //Log.Warning("Damage Info " + di.ToString() + " with " + di.damage + " initial damage");
+            if (victimBody && victimBody.master && victimBody.master.inventory && attackerBody)
+            {
+                //if (aCb.master.inventory){ Log.Debug("and has inventory"); }
+                int grabCount = MaxStack(victimBody.master.inventory);
+                if (grabCount > 0)
+                {
+                    Log.Debug(victimBody.GetUserName() + " takes " + damageReport.damageDealt + "\t silver thread damage with " + victimBody.healthComponent.fullCombinedHealth + "\t health");
+                    float deathChance = damageReport.damageDealt / victimBody.healthComponent.fullCombinedHealth * 100f * grabCount;
+                    if (deathChance > 100f)
+                    {
+                        deathChance = 100f;
+                    }
+                    Log.Debug("Chance of Snapping: " + deathChance);
+                    if (Util.CheckRoll(deathChance))
+                    {
+                        SnapBody(victimBody, attackerBody);
+                    }
+                }
+            }
+        }
+
+        // Thread snaps, Lose life or items
         private static void SnapBody(CharacterBody body, CharacterBody killer)
         {
             Log.Warning(body.GetUserName() + "'s thread was snapped by chance");
@@ -177,12 +180,13 @@ namespace UltitemsCyan.Items.Lunar
                 body.inventory.RemoveItemPermanent(lastItem.itemIndex, body.inventory.GetItemCountPermanent(lastItem));
                 body.inventory.RemoveItemTemp(lastItem.itemIndex, body.inventory.GetItemCountTemp(lastItem));
                 body.inventory.RemoveItemChanneled(lastItem.itemIndex, body.inventory.GetItemCountChanneled(lastItem));
-                //body.inventory.GiveItem(SilverThreadConsumed.item); // Don't actually give item but at least send notification of a broken item
+                // Give a Silver Threads for the stack broken
+                body.inventory.GiveItemPermanent(SilverThreadConsumed.item.itemIndex);
                 CharacterMasterNotificationQueue.SendTransformNotification(
                     body.master,
-                    item.itemIndex,
+                    lastItem.itemIndex,
                     SilverThreadConsumed.item.itemIndex,
-                    CharacterMasterNotificationQueue.TransformationType.Default);
+                    CharacterMasterNotificationQueue.TransformationType.None);
             }
             else
             {
@@ -318,99 +322,5 @@ namespace UltitemsCyan.Items.Lunar
             }
         }
 
-        // Make Scrapper return fewer items per Silver Thread Held
-        // ******************** OLD OLD OLD CODE CODE CODE ********************
-        /*
-        private void ScrapperController_BeginScrapping(On.RoR2.ScrapperController.orig_BeginScrapping orig, ScrapperController self, int intPickupIndex)
-        {
-            Log.Warning("Silver Scrapping check");
-            bool runOrig = true;
-            if (NetworkServer.active && self)
-            {
-                CharacterBody player = self.interactor.GetComponent<CharacterBody>();
-                if (player && player.master.inventory)
-                {
-                    //int grabSilverCount = MaxStack(player.master.inventory);
-                    if (player.master.inventory.GetItemCountEffective(item) > 0)
-                    {
-                        //Log.Debug("Silver Scrapping custom function");
-                        // body has a silver thread in their inventory
-                        runOrig = false;
-
-                        //self.itemsEaten = 0;
-                        PickupDef pickupDef = PickupCatalog.GetPickupDef(new PickupIndex(intPickupIndex));
-                        if (pickupDef != null && self.interactor)
-                        {
-                            //self.lastScrappedItemIndex = pickupDef.itemIndex;
-                            int scrapCount = Mathf.Min(self.maxItemsToScrapAtATime * costMultiplier, player.inventory.GetItemCountPermanent(pickupDef.itemIndex));
-                            //Log.Debug("Scrap Count: " + scrapCount);
-                            //Log.Debug(player.master.inventory.GetItemCount((ItemIndex)intPickupIndex) + " =? " + player.inventory.GetItemCount(pickupDef.itemIndex));
-                            if (scrapCount < costMultiplier)
-                            {
-                                // not enough items to convert item, don't return anything
-                                //Log.Debug("Silver Scrapper Consume poor items");
-                                //self.itemsEaten = -1;
-                            }
-                            else
-                            {
-                                // return reduced amount
-                                //Log.Debug("scrapCount: " + scrapCount + " returnCount: " + scrapCount / costMultiplier);
-                                player.inventory.RemoveItem(pickupDef.itemIndex, scrapCount);
-                                //self.itemsEaten += scrapCount / costMultiplier;
-                                for (int i = 0; i < scrapCount; i++)
-                                {
-                                    ScrapperController.CreateItemTakenOrb(player.corePosition, self.gameObject, pickupDef.itemIndex);
-                                }
-                                if (self.esm)
-                                {
-                                    self.esm.SetNextState(new EntityStates.Scrapper.WaitToBeginScrapping());
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-            // If checks failed, run original function
-            if (runOrig)
-            {
-                //Log.Debug("runOrig in SilverThread");
-                orig(self, intPickupIndex);
-                //Log.Debug("runOrig out SilverThread");
-            }
-        }
-        */
-
-        // Increase Items gained when given
-        public void Inventory_GiveItem_ItemIndex_int(On.RoR2.Inventory.orig_GiveItem_ItemIndex_int orig, Inventory self, ItemIndex itemIndex, int count)
-        {
-            //Log.Debug("SilverThread please start give item");
-            // && !inSilverAlready
-            if (!ItemCatalog.GetItemDef(itemIndex))
-            {
-                Log.Debug("SilverThread found impossible item? Index: " + itemIndex);
-            }
-            if (NetworkServer.active && count == 1 && self && ItemCatalog.GetItemDef(itemIndex).tier != ItemTier.NoTier && itemIndex != item.itemIndex)
-            {
-                // Precaution incase something causes an infinity loop of items
-                //inSilverAlready = true;
-                //Log.Debug("Do you have silver?");
-                int grabCount = MaxStack(self);
-                if (grabCount > 0)
-                {
-                    //Log.Debug("yes I do");
-                    //Log.Debug("Thread Chance: " + (baseThreadChance + (stackThreadChance * (grabCount - 1))));
-                    if (Util.CheckRoll(baseThreadChance + stackThreadChance * (grabCount - 1)))
-                    {
-                        Log.Debug("Extra silver: " + ItemCatalog.GetItemDef(itemIndex).name);
-                        // TODO add effect
-                        count += extraItemAmount;
-                    }
-
-                }
-            }
-            //Log.Debug("GiveItem_ItemIndex in orig SilverThread");
-            orig(self, itemIndex, count);
-            //Log.Debug("GiveItem_ItemIndex out orig SilverThread");
-        }
     }
 }