+++ UltitemsCyan/Items/Tier3/Grapevine.cs
@@ -1,5 +1,6 @@
 ﻿using RoR2;
 using UnityEngine;
+using static System.Math;
 using UnityEngine.AddressableAssets;
 using UnityEngine.Networking;
 using BepInEx.Configuration;
@@ -14,15 +15,18 @@ namespace UltitemsCyan.Items.Tier3
         private readonly static GameObject GrapeOrbPrefab = Addressables.LoadAssetAsync<GameObject>("RoR2/Base/Tooth/HealPack.prefab").WaitForCompletion();
         private readonly static GameObject GrapeEffect = Addressables.LoadAssetAsync<GameObject>("RoR2/Base/Common/VFX/HealthOrbEffect.prefab").WaitForCompletion();
 
-        private const float baseGrapeDropChance = 50f;
-        private const float stackGrapeDropChance = 25f;
+        //Play_acid_larva_impact
+
+        private const float baseGrapeDropChance = 40f;
+        //private const float stackGrapeDropChance = 25f;
 
         // For Slippery Buff
-        public const float grapeBlockChance = 80f;
+        public const float grapeBlockChance = 70f;
         public const int maxGrapes = 20;
 
         public override void Init(ConfigFile configs)
         {
+            Log.Warning("-JYPrint Hello?!?! 3");
             const string itemName = "Grapevine";
             if (!CheckItemEnabledConfig(itemName, "Red", configs))
             {
@@ -32,7 +36,7 @@ namespace UltitemsCyan.Items.Tier3
                 "GRAPEVINE",
                 itemName,
                 "Chance on kill to drop grapes that block damage.",
-                "<style=cIsHealing>50%</style> <style=cStack>(+25% per stack)</style> chance on kill to grow a grape. <style=cIsHealing>85%</style> to <style=cIsHealing>block</style> incomming damage per grape. Block chance is <style=cIsUtility>unaffected by luck</style>.",
+                "<style=cIsHealing>40%</style> chance on kill to grow <style=cIsHealing>1</style> <style=cStack>(+1 per stack)</style> grape. <style=cIsHealing>70%</style> to <style=cIsHealing>block</style> incoming damage per grape. Block chance is <style=cIsUtility>unaffected by luck</style>.",
                 "If you close your eyes, you can pretend their eyeballs",
                 ItemTier.Tier3,
                 UltAssets.GrapevineSprite,
@@ -43,7 +47,7 @@ namespace UltitemsCyan.Items.Tier3
 
         protected override void Hooks()
         {
-            On.RoR2.GlobalEventManager.OnCharacterDeath += GlobalEventManager_OnCharacterDeath; ;
+            On.RoR2.GlobalEventManager.OnCharacterDeath += GlobalEventManager_OnCharacterDeath;
         }
 
         private void GlobalEventManager_OnCharacterDeath(On.RoR2.GlobalEventManager.orig_OnCharacterDeath orig, GlobalEventManager self, DamageReport damageReport)
@@ -57,19 +61,17 @@ namespace UltitemsCyan.Items.Tier3
                 //int buffCount = killer.GetBuffCount(Buffs.OverclockedBuff.buff);
                 if (grabCount > 0)
                 {
-                    if (Util.CheckRoll(baseGrapeDropChance + (stackGrapeDropChance - 1) * grabCount, killer.master.luck))
+                    if (Util.CheckRoll(baseGrapeDropChance, killer.master.luck))
                     {
                         //Log.Warning("Dropping grape from " + victim.name);
-                        //RoR2.BuffPickup.Instantiate(item);
-                        //Util.PlaySound("Play_hermitCrab_idle_VO", victim.gameObject);
-                        //Util.PlaySound("Play_hermitCrab_idle_VO", victim.gameObject);
-                        SpawnOrb(victim.transform.position, victim.transform.rotation, TeamComponent.GetObjectTeam(killer.gameObject));
+                        Util.PlaySound("Play_clayGrenadier_impact", victim.gameObject);
+                        SpawnOrb(victim.transform.position, victim.transform.rotation, TeamComponent.GetObjectTeam(killer.gameObject), grabCount);
                     }
                 }
             }
         }
 
-        public static void SpawnOrb(Vector3 position, Quaternion rotation, TeamIndex teamIndex)
+        public static void SpawnOrb(Vector3 position, Quaternion rotation, TeamIndex teamIndex, int itemCount)
         {
             GameObject orb = Object.Instantiate(GrapeOrbPrefab);
             if (orb)
@@ -88,13 +90,13 @@ namespace UltitemsCyan.Items.Tier3
             trajectory.directionMode = VelocityRandomOnStart.DirectionMode.Cone;
             trajectory.coneAngle = 1;
 
-
             HealthPickup healthComponent = orb.GetComponentInChildren<HealthPickup>();
             //Log.Debug("health Component? " + healthComponent.alive);
             healthComponent.alive = false;
 
             //BuffPickup
             GrapePickup GrapeComponent = healthComponent.gameObject.AddComponent<GrapePickup>();
+            GrapeComponent.amount = itemCount;
             GrapeComponent.baseObject = orb;
             GrapeComponent.teamFilter = orb.GetComponent<TeamFilter>();
             GrapeComponent.pickupEffect = GrapeEffect;
@@ -119,16 +121,24 @@ namespace UltitemsCyan.Items.Tier3
             if (NetworkServer.active && alive && TeamComponent.GetObjectTeam(other.gameObject) == teamFilter.teamIndex)
             {
                 CharacterBody body = other.GetComponent<CharacterBody>();
-                if (body && body.GetBuffCount(buffDef) < maxGrapes)
+                //if (body && body.GetBuffCount(buffDef) < maxGrapes)
+                //{
+                //    body.AddBuff(buffDef);
+                //}
+                int addBuffs = Min(amount, maxGrapes - body.GetBuffCount(buffDef));
+                Log.Debug("Grape On Trigger Happened! amount: " + addBuffs);
+                for (int i = 0; i < addBuffs; i++)
                 {
                     body.AddBuff(buffDef);
                 }
+                _ = Util.PlaySound("Play_acid_larva_impact", body.gameObject);
                 Destroy(baseObject);
             }
         }
 
         //private BuffDef buffDef = JunkContent.Buffs.BodyArmor;
         private readonly BuffDef buffDef = Buffs.SlipperyGrapeBuff.buff;
+        public int amount;
 
         public GameObject baseObject;
         public TeamFilter teamFilter;